Replace uint8_t with uint16_t for AQI overflow resulting in negative values

+++ esphome/components/pmsa003i/pmsa003i.cpp
@@ -81,7 +81,7 @@ void PMSA003IComponent::update() {
     }
     */
 
-    int8_t aqi_value = -1;
+    uint16_t aqi_value = -1;
     if (this->aqi_sensor_ != nullptr && pm_2_5_value != -1 && pm_10_0_value != -1) {
       AbstractAQICalculator *calculator = this->aqi_calculator_factory_.get_calculator(this->aqi_calc_type_);
       aqi_value = calculator->get_aqi(pm_2_5_value, pm_10_0_value);
